requirements.txt
Flask==3.0.0
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-RESTX==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
stripe==9.4.0
redis==5.0.1
gunicorn==21.2.0
gevent==23.9.1
firebase-admin==6.5.0
